fix: Show dialog only when submitting an order
