+++ mobile/lib/features/stable/stable_dialog.dart
@@ -5,55 +5,46 @@ import 'package:get_10101/features/trade/submit_order_change_notifier.dart';
 import 'package:provider/provider.dart';
 
 class StableDialog extends StatelessWidget {
-  final PendingOrder pendingOrder;
-
-  const StableDialog({super.key, required this.pendingOrder});
+  const StableDialog({super.key});
 
   @override
   Widget build(BuildContext context) {
-    return Selector<SubmitOrderChangeNotifier, PendingOrderState>(
-      selector: (_, provider) => provider.pendingOrder!.state,
-      builder: (context, state, child) {
-        Widget body = createSubmitWidget(pendingOrder, DefaultTextStyle.of(context).style);
+    final submitOrderChangeNotifier = context.watch<SubmitOrderChangeNotifier>();
+    final pendingOrder = submitOrderChangeNotifier.pendingOrder!;
 
-        switch (state) {
-          case PendingOrderState.submitting:
-            return StableSubmissionStatusDialog(
-                title: pendingOrder.positionAction == PositionAction.open
-                    ? "Stabilizing"
-                    : "Bitcoinizing",
-                type: StableSubmissionStatusDialogType.pendingSubmit,
-                content: body);
-          case PendingOrderState.submittedSuccessfully:
-            return StableSubmissionStatusDialog(
-                title:
-                    pendingOrder.positionAction == PositionAction.open ? "Stabilize" : "Bitcoinize",
-                type: StableSubmissionStatusDialogType.successfulSubmit,
-                content: body);
-          case PendingOrderState.submissionFailed:
-            // TODO: This failure case has to be handled differently; are we planning to show orders that failed to submit in the order history?
-            return StableSubmissionStatusDialog(
-                title: pendingOrder.positionAction == PositionAction.open
-                    ? "Stabilizing"
-                    : "Bitcoinizing",
-                type: StableSubmissionStatusDialogType.failedSubmit,
-                content: body);
-          case PendingOrderState.orderFilled:
-            return StableSubmissionStatusDialog(
-                title:
-                    pendingOrder.positionAction == PositionAction.open ? "Stabilize" : "Bitcoinize",
-                type: StableSubmissionStatusDialogType.filled,
-                content: body);
-          case PendingOrderState.orderFailed:
-            return StableSubmissionStatusDialog(
-                title: pendingOrder.positionAction == PositionAction.open
-                    ? "Stabilizing"
-                    : "Bitcoinizing",
-                type: StableSubmissionStatusDialogType.failedFill,
-                content: body);
-        }
-      },
-    );
+    Widget body = createSubmitWidget(pendingOrder, DefaultTextStyle.of(context).style);
+
+    switch (pendingOrder.state) {
+      case PendingOrderState.submitting:
+        return StableSubmissionStatusDialog(
+            title:
+                pendingOrder.positionAction == PositionAction.open ? "Stabilizing" : "Bitcoinizing",
+            type: StableSubmissionStatusDialogType.pendingSubmit,
+            content: body);
+      case PendingOrderState.submittedSuccessfully:
+        return StableSubmissionStatusDialog(
+            title: pendingOrder.positionAction == PositionAction.open ? "Stabilize" : "Bitcoinize",
+            type: StableSubmissionStatusDialogType.successfulSubmit,
+            content: body);
+      case PendingOrderState.submissionFailed:
+        // TODO: This failure case has to be handled differently; are we planning to show orders that failed to submit in the order history?
+        return StableSubmissionStatusDialog(
+            title:
+                pendingOrder.positionAction == PositionAction.open ? "Stabilizing" : "Bitcoinizing",
+            type: StableSubmissionStatusDialogType.failedSubmit,
+            content: body);
+      case PendingOrderState.orderFilled:
+        return StableSubmissionStatusDialog(
+            title: pendingOrder.positionAction == PositionAction.open ? "Stabilize" : "Bitcoinize",
+            type: StableSubmissionStatusDialogType.filled,
+            content: body);
+      case PendingOrderState.orderFailed:
+        return StableSubmissionStatusDialog(
+            title:
+                pendingOrder.positionAction == PositionAction.open ? "Stabilizing" : "Bitcoinizing",
+            type: StableSubmissionStatusDialogType.failedFill,
+            content: body);
+    }
   }
 }
